Refactored secondary controller
