+++ controller/index.ts
@@ -1,2 +1,3 @@
 export { CacheController } from './cache.controller';
 export { TempController } from './temp.controller';
+export { SlaveController } from './slave.controller';
