+++ controller/temp.controller.ts
@@ -1,7 +1,6 @@
-// const ds18b20 = require('ds18b20');
 import * as bunyan from 'bunyan';
 import * as ds18b20 from 'ds18b20';
-import { CacheController } from './cache.controller';
+import { CacheController } from './';
 
 const SENSOR_ID = process.env.SENSOR_ID || '10-0008032d5234';
 
@@ -12,7 +11,7 @@ export class TempController {
 
   constructor(log: bunyan, cache: CacheController) {
     this.log = log;
-    this.log.info('Creating cache controller');
+    this.log.info('Creating temp controller');
     this.cache = cache;
   }
